+++ Solutions/Microsoft Copilot/ReleaseNotes.md
@@ -1,3 +1,4 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History**                                                                                |
 |-------------|--------------------------------|---------------------------------------------------------------------------------------------------|
-| 3.0.0       | 10-09-2025                     | Releasing a new Microsoft Copilot connector solution in Content Hub and Data Connector gallery.   |
+| 3.0.1       | 12-12-2025                     | Update copilot table name from LLMActivity to CopilotActivity.       |
+| 3.0.0       | 10-09-2025                     | Releasing a new Microsoft Copilot connector solution in Content Hub and **Data Connector** gallery.   |